requirements.txt
Flask>=2.2
psycopg2-binary>=2.9
python-dotenv>=1.0
sqlparse>=0.5
google-generativeai>=0.6